backend/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
mysqlclient==2.0.3
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2020.5
sqlparse==0.4.1
